ChaosMagnet/requirements.txt
dearpygui
psutil
numpy
pynput
opencv-python
sounddevice
requests
pyzmq

